+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.1] - 2019.05.08
+- Added import of magento 2 cms pages and blocks to the full import - @toh82 (#235)
+- Added information about magento 2 cms pages and blocks import to the readme - @toh82 (#235)
+
 ## [1.9.0] - 2019.05.06
 - Changed location of magento1 platform js client. Moved from `src/platform/magento1/module` to [magento1-vsbridge-client](https://github.com/DivanteLtd/magento1-vsbridge-client) - @mtarld (#195)
 - Update Babel from 6 to 7 - @lukeromanowicz